requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.37.0
apscheduler==3.10.4
pytz==2024.2
httpx==0.28.0
